+++ packages/nitrogen/src/views/CppHybridViewComponent.ts
