+++ packages/nitrogen/src/views/swift/SwiftHybridViewManager.ts
@@ -3,7 +3,7 @@ import type { HybridObjectSpec } from '../../syntax/HybridObjectSpec.js'
 import {
   createViewComponentShadowNodeFiles,
   getViewComponentNames,
-} from '../ViewComponentShadowNode.js'
+} from '../CppHybridViewComponent.js'
 import {
   createFileMetadataString,
   escapeCppName,
